test-requirements.txt
odoo-addon-hr_holidays_remaining_leaves @ git+https://github.com/OCA/hr-holidays.git@refs/pull/127/head#subdirectory=hr_holidays_remaining_leaves
